+++ src/main/java/com/paycal/api/PayCalView.java
@@ -1,12 +1,14 @@
 package com.paycal.api;
 
+import java.math.BigDecimal;
+
 /**
  * Created by edwin.njeru on 09/08/2017.
  */
 public interface PayCalView {
-    void displayResults(Double total,
-                        Double vatWithheld,
-                        Double withholdingTax,
-                        Double toPrepay,
-                        Double toPayee);
+    void displayResults(BigDecimal total,
+                        BigDecimal vatWithheld,
+                        BigDecimal withholdingTax,
+                        BigDecimal toPrepay,
+                        BigDecimal toPayee);
 }
